filtering by types of location and rating is done
